Design improved and routes added (No working yet)
Design of the navbar were improved with the colors decided, icons were changed.
Also, the 'a' classes were changed to 'Link' for routes purposes. In the Components were added the initials archives for routes purposes. And lately App.jsx were changed for the routes purposes. Was provided some clean to the code also.